new analyses of graph expressions  - truth prob analyse for graph expression    - input randomization method

+++ src/main/java/net/tislib/binanalyst/lib/calc/graph/tools/GraphCalculatorTools.java
@@ -145,7 +145,7 @@ public class GraphCalculatorTools {
         return calculator;
     }
 
-    public static Bit[] getTruthBit(BitOpsGraphCalculator calculator, Bit[] bits, int value) {
+    public static Bit[] getTruthBit(BitOpsGraphCalculator calculator, Bit[] bits, long value) {
         VarBit[] xBits = VarBit.list("x", bits.length, ZERO);
 
         setVal(xBits, value);
